+++ casanova-direct-links/casanova_direct_links.meta.js
@@ -1,16 +1,16 @@
 // ==UserScript==
-// @name           Batoto Direct Links
+// @name           Casanova Direct Links
 // @namespace      https://github.com/ToostInc/userscripts
 // @description    Adds Direct Links to the online reader
-// @include        http://www.batoto.net/read/_/*
+// @include        http://casanovascans.com/read/*
 // @author         Joost Bremmer < toost dot b at gmail dot com >
 // @copyright      2014, Joost Bremmer
 // @license        MIT
-// @version        2.0
+// @version        1.2
 // @date           06-08-2014
 // @require        http://code.jquery.com/jquery-latest.min.js
 // @grant          GM_addStyle
 // @grant          GM_xmlhttpRequest
-// @downloadURL    https://rawgit.com/ToostInc/userscripts/master/batoto-direct-links/batoto_direct_links.user.js
-// @updateURL      https://rawgit.com/ToostInc/userscripts/master/batoto-direct-links/batoto_direct_links.meta.js 
+// @downloadURL    https://rawgit.com/ToostInc/userscripts/master/casanova-direct-links/casanova_direct_links.user.js
+// @updateURL      https://rawgit.com/ToostInc/userscripts/master/casanova-direct-links/casanova_direct_links.meta.js 
 // ==/UserScript==